Replace Address and Hash types w/ byte vals
This removes the dependency on any go-ethereum repository,
forcing developers using the go bindings to explicitly cast
byte vals to whatever types are useful for their implementation.

Note that Address and Hash byte vals are declared with explicit
lengths and capacities.

This is a backwards incompatible change and so requires
conscientious version handling.